requirements.txt
aiogram~=3.18.0
pip~=25.0.1
attrs~=25.1.0
wheel~=0.43.0
idna~=3.10
multidict~=6.1.0
propcache~=0.2.1
pydantic~=2.10.6
aiofiles~=24.1.0
certifi~=2025.1.31
aiohttp~=3.11.12
yarl~=1.18.3
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.6
setuptools~=75.8.0
packaging~=24.2
tomli~=2.0.1
autocommand~=2.2.2
inflect~=7.3.1
typeguard~=4.3.0
platformdirs~=4.2.2
zipp~=3.19.2
python-dotenv~=1.0.1